Moved CsvLine class to CsvManager.cs file
